Edit index.js and move code awesomebook class into seperate module
